requirements.txt
VersaLog
discord
line-bot-sdk
python-dotenv
Flask